Refactor daemon and semanage code into separte packages
This is meant to enable easier testing as well as more defined
interfaces.

The interface that manages the semodule interactions is now its own
thing:

    `SEModuleHandler`

And now there are two implementations of it:

* `SemanageHandler`: Is the implementation we had before. This opens a
  connection with semanage, which is closed with the standard `Close`
  interface

* `SEModuleTestHandler`: Is a test implementation of a handler. This
aids in us unit-testing this code.

The Daemon is in it's own package as well and now only takes into use
the aforementioned interface.