Provide a UNIX socket that allows listing policies
adds a new goroutine to the deamon that creates a socket that the deamon
listens on. There is an HTTP server listening on the socket that handles
a single path /policies and only handles HTTP GET requests. The answer
to that request is a JSON list of all the loaded policies as string,
similar to what a call to semanage -l would produce.

There is so far no access control on the socket except for the UNIX
permissions of the socket which are settable through command line
parameters.