semanage/getNthModName: Free modInfo inside function
When removing/uninstalling a module, we first check if the module is
available in the current priority. This is done by listing all the
modules (modInfoList), getting the information of each one (modInfo),
and checking if the module is in that list.

We would normally free the modInfo as part of a defer call... but this
actually gets executed after the call to list. This means that all the
modInfo objects are loaded in memory until they're all free'd once the
List function ends.

This PR changes that behavior by calling `free` inside the function that
gets the info for the module. Thus alleviating this memory spike.