Implement policy datastore
This is an initial implementation of the policy datastore.

The datastore is meant to store policies that are installed
successfully and unsuccesfully throughout the lifetime of selinuxd.
It is based on bbolt which is a lightweight key/value store [1].

In order to use the datastore, an interface called DataStore and
ReadOnlyDataStore are available. This allows us to pass a reference to
the datastore without depending too much on bbolt's interfaces.

Getting, Removing and Listing of policies is implemented. Whenever a
policy is installed, it is persisted in the datastore. The an equivalent
operation happens when an explicit removal happens.

The status server now uses the datastore instead of the semanage handler.

TODO
====

* Remove missing policies on startup
* Add the ability to query specific policies and their status in the
  server.

[1] https://github.com/etcd-io/bbolt/

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>