Implement policy-specific status in status server
This adds the ability for the status server to output policy-specific
statuses.

The path is a simple GET request to `/policies/{policy name}`.

This introduced a global buffer that captures the error strings coming
from the libsemanage library. This gets flushed with every commit and is
only taken into account if there was an error committing a policy.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>