Expose AutoCommit and do one commit (best-case) in oneshot
This exposes the Commit operation in the SeModule Handler code. It also
exposes the auto commit setting. What this means is that now we're able
to tell the library that we want to gather all modules and try to
install them all as part of the same transaction as opposed to doing it
per policy.

This is then used in the oneshot command, where the command will try to
do a single transaction to install all the policies, and if that fails
it'll go policy per policy installing them. This is meant to speed up
setting up initial policies.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>