Implement checksumming of policies
This gets a checksum for the policy before it's installed.

If the policy is already in the datastore, and the checksum matches, the
installation step is skipped (thus making it faster).

If it's not there, it actually gets installed.

note that the checksum will not be displayed in the status at all.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>