Add a new semodule back end that forks out to policycoreutils to avoid memory leaks
During stress-testing of SPO it was found that selinuxd leaks memory. It
appears that the leak is coming from libsemanage (confirmed with
valgrind and fixed upstream). Using the policycoreutils back end,
testing on OpenShift 4.x, the selinuxd containers spike at around ~400MB
before falling back to ~30-50 MB. With the semanage back end, the memory
would climb steadily until it would hit the 1GB limit imposed by SPO's
DS definition and be OOMKilled.

To work around that, this PR adds a new back-end for the semodule
package that shells out to the policycoreutils utilities. This back end
is the default on "RHEL-like" distributions, but otherwise selectable
during build using the SEMODULE_BACKEND environment variable that
translates into a build tag.

The "Handler" interface was moved from semodule top-level to
semodule/seiface to avoid circular dependencies.

Signed-off-by: Jakub Hrozek <jhrozek@redhat.com>