Use priority 350 for selinuxd modules
SELinux modules store has a concept of priorities which can be used for
detection of the source of modules:

    # semodule -lfull
    400 agent-udica       cil
    200 cockpit           pp
    200 container         pp
    ...
    100 abrt              pp
    100 accountsd         pp
    ...

In Red Hat Enterprise Linux and Fedora we use priority 100 for modules
from selinux-policy-targeted package, 200 is used for modules from
packages shipping their own policies, 300 is suggested by
SETroubleshoot, and 400 is the default for user modules.

Priority 350 for selinuxd modules was suggested by SELinux team.

Signed-off-by: Petr Lautrbach <plautrba@redhat.com>