Add release Makefile targets
Adds several Makefile targets to help create stable releases. The
top-level one is "make releas" which should be called with a
RELEASE_VERSION variable, e.g.:

make release RELEASE_VERSION=0.3.0

Doing that changes the VERSION file, commits and tags that commit with
the release version. Then, images using both Fedora and CentOS are built
using that version as the tag instead of latest and pushed, finally the
same images are built and pushed also using :latest.

Signed-off-by: Jakub Hrozek <jhrozek@redhat.com>