Use `semodule -lfull --checksum` instead of the datastore
SELinux userspace release 3.4 introduced a new command line option
[-m|--checksum] to `semodule` which adds sha256 checksum of modules to
its output. It can be used to check whether the same module is already
installed or not. Given that selinuxd installed modules use priority 350
we can  use semodule checksum and priority 350 as an indicator whether a
module was already installed by selinuxd or not and therefore there's no
need to track the state of modules in a separate datastore.

`semodule --checksum` is supported since Red Hat Enterprise Linux 8.6

Signed-off-by: Petr Lautrbach <lautrbach@redhat.com>